feat(common): 合并通用专辑详情面板组件
- 将原有的横/竖屏专用组件合并为通用的专辑详情面板组件(AlbumSection)。
- 修改使用自适应设计，根据分配的空间动态调整布局和尺寸。
